object selection support by Mark Borg
In brief, the change that I did is to allow the user to switch back and forth between object selection/initialisation stage and tracking stage via a key press. In object selection stage, the user can use the mouse to left-button click and drag the mouse to define the bounding box. Added also OpenCV-style command-line parameters to start tracking automatically